requirements/hyena.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
biopython==1.86
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.3.1
cmake==4.2.0
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
datasets==4.4.1
debugpy==1.8.18
decorator==5.2.1
dill==0.4.0
einops==0.8.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.19.1
flash-attn @ https://github.com/Dao-AILab/flash-attention/releases/download/v2.5.6/flash_attn-2.5.6+cu118torch2.1cxx11abiFALSE-cp311-cp311-linux_x86_64.whl
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.9.0
gdown==5.2.0
genomic-benchmarks==1.0.0
git-lfs==1.6
gitdb==4.0.12
gitpython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.4
ipykernel==7.1.0
ipython==9.8.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.7.0
jupyter-core==5.9.1
jupyterlab-widgets==3.0.16
kiwisolver==1.4.9
liftover==1.3.3
lightning-utilities==0.15.2
loguru==0.7.3
markdown-it-py==4.0.0
markupsafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
narwhals==2.13.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
numerize==0.12
numpy==1.26.4
omegaconf==2.3.0
opt-einsum==3.4.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
pip==25.3
platformdirs==4.5.1
plotly==6.5.0
polars==1.36.1
polars-runtime-32==1.36.1
prettytable==3.17.0
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==22.0.0
pydantic==2.12.5
pydantic-core==2.41.5
pyfaidx==0.9.0.3
pygments==2.19.2
pyparsing==3.2.5
pysocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==1.8.6
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
sentry-sdk==2.47.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
tensorboardx==2.6.4
threadpoolctl==3.6.0
timm==1.0.22
tokenizers==0.13.3
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchmetrics==1.8.2
torchvision==0.16.0+cu118
tornado==6.5.3
tqdm==4.67.1
traitlets==5.14.3
transformers==4.26.1
triton==2.1.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==1.26.13
wandb==0.23.1
wcwidth==0.2.14
wheel==0.45.1
widgetsnbextension==4.0.15
xxhash==3.6.0
yarl==1.22.0
